MolecularGraphKernels.jl

Graph kernels transform a graph $G$ into a fixed-size vector $\bold{u}$. This is useful, for example, when one wishes to train a machine learning model on graph-structured inputs of varying size, topology, and node/edge metadata.

MetaGraphs.MetaGraph — Method

g = MetaGraph(mol)

Convert a GraphMol object into the corresponding MetaGraph

Note

Hydrogen atoms are generally treated implicitly.

MolecularGraphKernels.banner — Method

RWGKSVM.banner()

Prints the stylized ASCII console banner for the package.

MolecularGraphKernels.csi_adj_mat — Method

cpg = csi_adj_mat(graph_a, graph_b)

Returns the adjacency matrix of the CSI product graph of graph_a and graph_b

MolecularGraphKernels.csi_product_graph — Method

cpg = csi_product_graph(graph_a, graph_b)

Returns the common subgraph isomorphism product graph of graph_a and graph_b

MolecularGraphKernels.direct_product_graph — Method

dpg = direct_product_graph(graph_a, graph_b)

Returns the direct product graph of graph_a and graph_b.

Note

If only the adjacency matrix of the DPG is needed (and not the DPG itself) @ref(dpg_adj_mat) should be used instead for performance.

Warning

Edge and vertex labels should be of consistent, comparable types, preferably Symbol, Int, Bool, etc., for efficiency. Comparing, for example, String labels will result in significantly poorer performance!

MolecularGraphKernels.dpg_adj_mat — Method

A = dpg_adj_mat(graph_a, graph_b)

Returns the adjacency matrix of the DPG.

MolecularGraphKernels.random_walk_kernel — Method

kernel_score = random_walk_kernel(adj_mat, l)
kernel_score = random_walk_kernel(dpg, l)
kernel_score = random_walk_kernel(A, B, l)

Returns the similarity score for two graphs by applying the l-length random walk graph kernel on their direct product graph.