Product Graphs

Product graphs are contained within the ProductGraph type. The ProductGraph type is parameterized by subtypes of AbstractProductGraph.

subtypes(AbstractProductGraph)

# output

3-element Vector{Any}:
 Direct
 Modular
 Weighted

While a ProductGraph contains the full representation of the product graph between two molecules, the adjacency matrix may be all that is required. As such, it may be more efficient to work with the adjacency matrix (see GraphMatrix).

To create a new ProductGraph, define a new AbstractProductGraph subtype and two functions, record_adjacency! and product_graph_edge_label, parameterized by the new type.

MolecularGraphKernels.ProductGraph — Type

type-parameterized struct for product graphs

Missing docstring.

Missing docstring for GraphMatrix. Check Documenter's build log for details.