MolecularGraphKernels.jl

Graph kernels transform a graph $G$ into a fixed-size vector $\bold{u}$. This is useful, for example, when one wishes to train a machine learning model on graph-structured inputs of varying size, topology, and node/edge metadata.

MetaGraphs.MetaGraph — Method

g = MetaGraph(mol)

Convert a GraphMol object into the corresponding MetaGraph

MolecularGraphKernels.banner — Method

RWGKSVM.banner()

Prints the stylized ASCII console banner for the package.

MolecularGraphKernels.direct_product_graph — Method

dpg_adj_mat = direct_product_graph(graph_a, graph_b)

Returns the adjacency matrix of the direct product graph of graph_a and graph_b.

dpg_adj_mat = direct_product_graph(graph_a, graph_b; return_graph=true)

Returns the adjacency matrix and the

MolecularGraphKernels.dpg — Method

g = dpg(mat, map, mol)

Returns the direct product graph based on its adjacency matrix, the map of A/B nodes to DPG nodes, and graph A

MolecularGraphKernels.fixed_length_rw_kernel — Method

kernel_score = fixed_length_rw_kernel(adj_mat, l)
kernel_score = fixed_length_rw_kernel(dpg, l)
kernel_score = fixed_length_rw_kernel(A, B, l)

Returns the similarity score for two graphs by applying the l-length random walk graph kernel on their direct product graph.