MolecularGraphKernels.jl

Graph kernels transform a graph $G$ into a fixed-size vector $\bold{u}$. This is useful, for example, when one wishes to train a machine learning model on graph-structured inputs of varying size, topology, and node/edge metadata.

MetaGraphs.MetaGraph — Method

g = MetaGraph(mol)

Convert a GraphMol object into the corresponding MetaGraph

Note

Hydrogen atoms are generally treated implicitly.

MolecularGraphKernels.banner — Method

RWGKSVM.banner()

Prints the stylized ASCII console banner for the package.

MolecularGraphKernels.build_v₁v₂_pair_to_w_map — Method

map (v₁ ∈ g₁, v₂ ∈ g₂) ↦ w ∈ g₁ x g₂

MolecularGraphKernels.build_w_to_v₁v₂_pair_map — Method

map w ∈ g₁ x g₂ ↦ (v₁ ∈ g₁, v₂ ∈ g₂)

MolecularGraphKernels.graph_kernel — Method

kernel_score = graph_kernel(adj_mat, l)
kernel_score = graph_kernel(g₁xg₂, l)
kernel_score = graph_kernel(g₁, g₂, l, type)

Returns the similarity score for two graphs by applying the l-length random walk graph kernel on their type product graph.

MolecularGraphKernels.product_graph — Method

compute the product graph (of type type) between g₁ and g₂

MolecularGraphKernels.product_graph_adj_mat — Method

compute the adjacency matrix of the product graph (of type type) between g₁ and g₂, but do not explicitly construct the graph

MolecularGraphKernels.product_graph_matrix_and_maps — Method

compute the product graph adjacency matrix and mappings (in each direction) for relating the source graphs and product graph